add class_Category.py, class_Product.py, main.py fix test_class.py
