+++ src/class_Product.py
@@ -1,0 +1,13 @@
+class Product:
+    name : str
+    description : str
+    price : float
+    quantity : int
+    def __init__(self,name ,description ,price ,quantity):
+        self.name = name
+        self.description = description
+        self.price = price
+        self.quantity =quantity
+
+    def __repr__(self):
+        return self.name
